refactor: update page imports after filter system relocation (#245)
