+++ frontend/src/pages/sales/PaymentsPage.tsx
@@ -44,16 +44,17 @@ import {
 import { formatCurrency, formatDate, formatWholeQuantity } from '@/utils/formatters'
 import { TABLE_STYLES } from '@/constants/tableStyles'
 import { ListSkeleton } from '@/components/common/ListSkeleton'
-import { FilterBar, useFilterBar } from '@/components/filters'
-import type { FilterBarConfig } from '@/components/filters'
+import { FilterBar } from '@/components/filters'
 import DeletedPaymentsDialog from '@/components/sales/DeletedPaymentsDialog'
 import PageHeader from '@/components/common/PageHeader'
 import { PaymentReceiptPrint } from '@/components/print'
+import { useFilterBar } from '@/hooks/useFilterBar'
 import { useKeyboardShortcuts } from '@/hooks/useSearchAndFilter'
 import { useLazyGetJournalEntriesQuery } from '@/store/api/accountingApi'
 import { useAppDispatch, useAppSelector } from '@/hooks/useRedux'
 import { useGetCustomersQuery, useGetPaymentsQuery } from '@/store/api/salesApi'
 import { setSelectedPayment, selectSelectedPayment } from '@/store/slices/salesSlice'
+import type { FilterBarConfig } from '@/types/filterBar.types'
 import type { InvoiceItem } from '@/types'
 
 // Payment types and interfaces
